Feat: 채팅오류수정
